+++ src/server/game/Entities/Unit/Unit.cpp
@@ -2427,7 +2427,7 @@ void Unit::DealMeleeDamage(CalcDamageInfo* damageInfo, bool durabilityLoss)
 
     // If this is a creature and it attacks from behind it has a probability to daze it's victim
     if ((damageInfo->hitOutCome == MELEE_HIT_CRIT || damageInfo->hitOutCome == MELEE_HIT_CRUSHING || damageInfo->hitOutCome == MELEE_HIT_NORMAL || damageInfo->hitOutCome == MELEE_HIT_GLANCING)
-        && !IsPlayer() && !ToCreature()->IsControlledByPlayer() && !victim->HasInArc(M_PI, this)
+        && !IsPlayer() && !ToCreature()->IsControlledByPlayer() && !victim->HasInArc(float(M_PI), this)
         && (victim->IsPlayer() || !victim->ToCreature()->isWorldBoss()) && !isTank)
     {
         // -probability is between 0% and 40%
@@ -3387,7 +3387,7 @@ bool Unit::isSpellBlocked(Unit* victim, SpellInfo const* spellProto, WeaponAttac
     if (victim->IsNonMeleeSpellCast(false) || victim->HasUnitState(UNIT_STATE_CONTROLLED))
         return false;
 
-    if (victim->HasAuraType(SPELL_AURA_IGNORE_HIT_DIRECTION) || victim->HasInArc(M_PI, this))
+    if (victim->HasAuraType(SPELL_AURA_IGNORE_HIT_DIRECTION) || victim->HasInArc(float(M_PI), this))
     {
         // Check creatures flags_extra for disable block
         if (victim->IsCreature() && victim->ToCreature()->GetCreatureTemplate()->flags_extra & CREATURE_FLAG_EXTRA_NO_BLOCK)
@@ -11997,7 +11997,7 @@ void Unit::SetMinion(Minion *minion, bool apply)
         uint32 count = 0;
         for (ControlList::iterator itr = m_Controlled.begin(); itr != m_Controlled.end(); ++itr)
         {
-            float angle = PET_FOLLOW_ANGLE;
+            float angle = float(PET_FOLLOW_ANGLE);
             Creature* creature = ObjectAccessor::GetCreatureOrPetOrVehicle(*this, *itr);
             if (!creature)
                 continue;
@@ -12005,23 +12005,23 @@ void Unit::SetMinion(Minion *minion, bool apply)
             switch (count)
             {
                 case 0:
-                    angle = PET_FOLLOW_ANGLE;
+                    angle = float(PET_FOLLOW_ANGLE);
                     break;
                 case 1:
-                    angle = PET_FOLLOW_ANGLE * 3;
+                    angle = float(PET_FOLLOW_ANGLE * 3);
                     break;
                 case 2:
-                    angle = PET_FOLLOW_ANGLE * 2;
+                    angle = float(PET_FOLLOW_ANGLE * 2);
                     break;
                 case 3:
-                    angle = PET_FOLLOW_ANGLE * 2 + PET_FOLLOW_ANGLE / 2;
+                    angle = float(PET_FOLLOW_ANGLE * 2 + PET_FOLLOW_ANGLE / 2);
                     break;
                 case 4:
-                    angle = PET_FOLLOW_ANGLE + PET_FOLLOW_ANGLE / 2;
+                    angle = float(PET_FOLLOW_ANGLE + PET_FOLLOW_ANGLE / 2);
                     break;
             }
             creature->SetFollowAngle(angle);
-            creature->SetFollowDistance(frand(PET_FOLLOW_DIST/ 2 , PET_FOLLOW_DIST *2));
+            creature->SetFollowDistance(frand(PET_FOLLOW_DIST / 2 , PET_FOLLOW_DIST * 2));
             count++;
         }
     }
