+++ src/server/scripts/BrawlersGuild/brawlers_guild_bosses_rank_four.cpp
@@ -156,7 +156,7 @@ public:
                 {
                     case EVENT_1:
                     Position pos;
-                    me->GetNearPoint2D(pos, 15.0f, frand(-M_PI/3, M_PI/3));
+                    me->GetNearPoint2D(pos, 15.0f, frand(float(- M_PI / 3), float(M_PI / 3)));
                     me->CastSpell(pos.GetPositionX(), pos.GetPositionY(), pos.GetPositionZ(), SPELL_WATER_ORB);
                     events.RescheduleEvent(EVENT_1, urand (17000, 19000));
                     break;
