+++ src/server/scripts/BrawlersGuild/brawlers_guild_bosses_rank_seven.cpp
@@ -75,7 +75,7 @@ public:
                     case EVENT_2:
                     {
                         Position pos;
-                        me->GetNearPoint2D(pos, 15.0f, frand(-M_PI/3, M_PI/3));
+                        me->GetNearPoint2D(pos, 15.0f, frand(float(- M_PI / 3), float(M_PI / 3)));
                         me->CastSpell(pos.GetPositionX(), pos.GetPositionY(), pos.GetPositionZ(), SPELL_ACIT_AT);
                         events.RescheduleEvent(EVENT_2, 20000);
                         break;
