+++ src/server/scripts/Kalimdor/Firelands/boss_lord_rhyolith.cpp
@@ -388,7 +388,7 @@ class boss_lord_rhyolith : public CreatureScript
                                 std::list<Creature*> volcanos_1;
                                 for (std::list<Creature*>::const_iterator itr = volcanos.begin(); itr != volcanos.end(); ++itr)
                                 {
-                                    if (me->HasInArc(M_PI / 2, (*itr)))
+                                    if (me->HasInArc(float(M_PI / 2), (*itr)))
                                         volcanos_1.push_back((*itr));
                                 }
                                 pTarget = Trinity::Containers::SelectRandomContainerElement((volcanos_1.empty() ? volcanos : volcanos_1));
@@ -1257,7 +1257,7 @@ class spell_lord_rhyolith_magma_flow : public SpellScriptLoader
 
                 count++;
 
-                for (float a = 0; a <= 2 * M_PI; a += M_PI / 2)
+                for (float a = 0; a <= 2 * M_PI; a += float(M_PI / 2))
                 {
                     Position pos = GetCaster()->GetNearPosition(1.0f * count, a + frand(-0.05f, 0.05f));
                     GetCaster()->CastSpell(pos.GetPositionX(), pos.GetPositionY(), pos.GetPositionZ(), SPELL_MAGMA_FLOW_AREA, true);
