+++ src/server/scripts/Legion/HallsofValor/boss_god_king_skovald.cpp
@@ -252,7 +252,7 @@ public:
                         me->CastSpell(target, 193686, true);
                     break;
                 case SPELL_INFERNAL_FLAMES_FIND_N:
-                    if (!me->isInBack(target, M_PI * 0.8f))
+                    if (!me->isInBack(target, float(M_PI * 0.8f)))
                         DoCast(target, SPELL_INFERNAL_FLAMES_TRIG_AT, true);
                     break;
                 case SPELL_INFERNAL_FLAMES_FIND_H:
@@ -881,7 +881,7 @@ class spell_skovald_aegis_remove : public SpellScriptLoader
                     return;
 
                 if (Unit* attacker = dmgInfo.GetAttacker())
-                    if (attacker->GetDistance(GetCaster()) < 5.0f && GetCaster()->isInBack(attacker, M_PI * 0.8f))
+                    if (attacker->GetDistance(GetCaster()) < 5.0f && GetCaster()->isInBack(attacker, float(M_PI * 0.8f)))
                         absorbAmount = 0;
                     else
                         absorbAmount = dmgInfo.GetDamage();
@@ -925,8 +925,8 @@ class spell_skovald_aegis_absorb : public SpellScriptLoader
                 if (auto attacker = dmgInfo.GetAttacker())
                 {
                     Position pos;
-                    GetCaster()->SimplePosXYRelocationByAngle(pos, 5.0f, M_PI);
-                    if (pos.HasInArc(M_PI * 1.4f, attacker))
+                    GetCaster()->SimplePosXYRelocationByAngle(pos, 5.0f, float(M_PI));
+                    if (pos.HasInArc(float(M_PI * 1.4f), attacker))
                         absorbAmount = dmgInfo.GetDamage();
                     else
                         absorbAmount = 0;
